refactor: move CleanupOnTimeout to common package
