+++ internal/service/advancedclustertpf/common.go
@@ -1,8 +1,6 @@
 package advancedclustertpf
 
 import (
-	"context"
-	"errors"
 	"fmt"
 	"strings"
 	"time"
@@ -115,16 +113,4 @@ func GetPriorityOfFlexReplicationSpecs(replicationSpecs *[]admin.ReplicationSpec
 		return nil
 	}
 	return regionConfig.Priority
-}
-
-func CleanupOnTimeout(ctx context.Context, diags *diag.Diagnostics, warningDetail string, cleanup func(context.Context) error) {
-	if !errors.Is(ctx.Err(), context.DeadlineExceeded) {
-		return
-	}
-	cleanupWarning := "Failed to create, will perform cleanup due to timeout reached"
-	diags.AddWarning(cleanupWarning, warningDetail)
-	newContext := context.Background() // Create a new context for cleanup
-	if err := cleanup(newContext); err != nil {
-		diags.AddWarning("Error during cleanup", warningDetail+" error="+err.Error())
-	}
 }
